Add and make use of ECM clang-format integration
This formats the codebase according to KDE's coding style using clang-format.

To esure the style is consistent going forward this also adds a commit hook that
verfies that changes are properly formatted when committing.

This helps both contributors and reviewers to take away the mental work of paying
attention to the coding style.

This is common practice to do for many KDE projects