Working version of the freespace info
This code will just wait for all the KIO:Job to finish getting the free
space info and then it will continue.

This seems less than ideal, but calling updateList() freeSpaceResult
wastes a lot of space because the icons are updated all the time, which
takes a lot of time. Perhaps freeSpaceResult() could call a less
demanding version?